requirements.txt
# streamlit
# pypdf2
# langchain
# python-dotenv
# faiss-cpu
# huggingface_hub
# InstructorEmbedding 
# sentence-transformers==2.2.2
# #-e .
langchain==0.0.184
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1

faiss-cpu==1.7.4
#faiss-gpu

altair==4
tiktoken==0.4.0

huggingface-hub==0.14.1

InstructorEmbedding==1.0.1
sentence-transformers==2.2.2


# haystack.retriever
# haystack.document_store.faiss
# haystack.retriever
#farm-haystack[colab,faiss,inference]